### 2024-07-02 #6951 @editorialbot
[PRE REVIEW]: FAIRmaterials: Ontology Tools with Data FAIRification in Development
**Submitting author:** @Alexhb02 (<a href="http://orcid.org/0009-0000-2917-773X">Alexander Harding Bradley</a>)
**Repository:** https://github.com/cwru-sdle/FAIRmaterials
**Branch with paper.md** (empty if default branch): main
**Version:** 0.4.2.2
