### 2024-07-02 c7ae579
affiliation and doi additions
