+++ JOSSpaper/paper.bib
@@ -59,6 +59,7 @@ isbn="978-3-319-17966-7"
   author       = {Rich Iannone and Olivier Roy},
   title        = {{DiagrammeR: Graph/Network Visualization}},
   howpublished = {\url{https://cran.r-project.org/web/packages/DiagrammeR/index.html}},
+  doi          = {10.32614/CRAN.package.DiagrammeR},
   note         = {Accessed: 2024-06-18}
 }
 
@@ -83,6 +84,7 @@ isbn="978-3-319-17966-7"
   author       = {Carl Boettiger, Bryce Mecum, Anna Krystalli, Viktor SenderovD},
   title        = {{rdflib: Tools to Manipulate and Query Semantic Data}},
   howpublished = {\url{https://cran.r-project.org/web/packages/rdflib/index.html}},
+  doi          = {10.32614/CRAN.package.rdflib},
   note         = {Accessed: 2024-06-18}
 }
 
