### 2024-07-03 #6951 @Jonathan-E-Gordon
@danielskatz The software was published on CRAN and PyPI in 2021. Due to lab policy, we developed the software in a private Bitbucket repository, which has accumulated over 400 commits since 2021. In January 2024, we initiated a redesign of the software, resulting in 110 commits from 9 authors, representing approximately 7 person-months of work. We can send the whole commit history if that is helpful. This software will also be cited in a scientific data paper currently under review, titled “Materials Data Science Ontology (MDS-Onto): Unifying Domain Knowledge in Materials and Applied Data Science” and we believe this software will be very valuable for those generating ontologies, addressing a need in various academic domains.
