### 2024-07-09 #6951 @danielskatz
👋 @atrisovic - I know you have a bunch of JOSS papers now, but I thought I would see if you are interested in taking on this one, as it seem aligned with your interests. If not, I will probably take it myself, in which case I would appreciate any reviewer suggestions you might have.
