### 2024-07-09 #6951 @danielskatz
@Alexhb02 & @Jonathan-E-Gordon - as [previously mentioned](https://github.com/openjournals/joss-reviews/issues/6951#issuecomment-2204272019), JOSS papers are expected to be 250-1000 words in length. This paper is currently over 2500, as shown by the `@editorialbot check repository` command. If JOSS accepts this for review, you will have to reduce it. Additionally, the remaining suggested DOI appears correct, so at some point, you will need to add it to the entry in the .bib file.
