### 2024-07-09 #6951 @danielskatz
@Alexhb02 & @Jonathan-E-Gordon - as [previously mentioned](https://github.com/openjournals/joss-reviews/issues/6951#issuecomment-2204272019), JOSS papers are expected to be 250-1000 words in length. This paper is currently over 2500, as shown by the `@editorialbot check repository` command. You have to reduce it. Additionally, the remaining suggested DOI (10.1007/978-3-319-17966-7_21 may be a valid DOI for title: WebVOWL: Web-based Visualization of Ontologies) appears correct, so you need to add it to the entry in the .bib file.
