### 2024-07-09 #6951 @Jonathan-E-Gordon
@danielskatz Would a paper length of about 1200 words be acceptable or does the paper need to be under 1000 words?
