### 2024-07-09 #6951 @danielskatz
1200 is probably fine.
