+++ JOSSpaper/paper.md
@@ -7,9 +7,9 @@ tags:
 - FAIR
 date: "26 June 2024"
 output:
-  pdf_document: default
   html_document:
     df_print: paged
+  pdf_document: default
 authors:
 - name: Alexander Harding Bradley
   affiliation: 2, 3
@@ -54,7 +54,6 @@ affiliations:
 editor_options: 
   markdown: 
     wrap: sentence
-
 ---
 
 # Summary
@@ -62,12 +61,9 @@ editor_options:
 The bilingual `FAIRmaterials` package is a robust and user-friendly tool designed to simplify the creation and visualization of materials and data science ontologies.
 `FAIRmaterials`, available in the Python and R languages, addresses the complexities and technical barriers associated with traditional ontology editors based on manual user input such as `Protege` [@Protege].
 `FAIRmaterials` offers an intuitive workflow and easy-to-use templates, making it accessible to users both experienced and inexperienced with ontology creation.
-By leveraging a CSV-based template system, the package allows for the straightforward design and expansion of ontologies, promoting better data accessibility, management, and interoperability in research environments.
 
 The core feature of the `FAIRmaterials` package is its ability to programatically convert simple and structured CSV inputs into rich, well-defined ontologies.
 This capability is designed to support the findability, accessibility, interoperability, and reusability (FAIR) [@wilkinsonFAIRGuidingPrinciples2016] of research data and serve as a tool in the process of data FAIRification.
-`FAIRmaterials` also integrates seamlessly with other existing data systems and protocols within the scientific community.
-The package supports various serialization formats, such as Turtle and JSON-LD, to accommodate different user preferences and technical needs.
 
 Designed with the end-user in mind, `FAIRmaterials` facilitates a comprehensive yet entry-level experience in ontology creation.
 Its additional features, such as automated ontology merging, static visualizations, and comprehensive documentation for outputs extend its utility, making it a valuable tool for any researcher engaged in knowledge management.
@@ -82,39 +78,25 @@ Therefore, there is a need for a tool that can create ontologies with an interfa
 `FAIRmaterials` seeks to lower the barrier of entry for scientists entering the world of ontology development and evolution.
 The package provides a baseline CSV ontology template with built-in and easy-to-follow instructions on how to design an ontology which can be found [here](https://docs.google.com/spreadsheets/d/1k7gFRc1Yslc-m65aWfFCxqk5UtrKZl9c3GyEFQvLSFU/edit?gid=1078491905#gid=1078491905).
 
-The CSV format was chosen as an input for `FAIRmaterials` due to its well-established familiarity within the scientific community as well as its ease of use without the need for specialized tools or programs.
-The primary function of our package is to programmatically convert the filled-out CSV files into a well-structured ontology file.
-In addition to outputting the ontology in multiple syntaxes, the package outputs other files to aid in the ontology development and publishing process, such as a static visualization and an HTML documentation file about the ontology.
-
 ![Empty Variable CSV Template Sheet for the `FAIRmaterials` Package. The CSV template sheet includes specific instructions on how to fill out every row to correctly generate the ontology. Template is split in half for readability.\label{fig:Variable_Sheet_Template_CSV}](Variable_Sheet_Template_CSV.png){width="100%"}
 
 # Key Features
 
 ## Ontology creation from template CSVs
 
-The primary function of the FAIRMaterials package is to convert the term, relationship, and value specifications from the CSV template into an ontology.
-Both the Python and R versions of the package rely on an open-source library called `RDFLib` [@RDFLibR; @RDFLibPython], which allows users to programmatically create and populate an ontology, serialize it into many syntaxes, validate its logical consistency, and provide output documentation.
-The package parses through the CSV sheets and generates an `RDFLib` graph object of the ontology.
[…]
