+++ JOSSpaper/paper.bib
@@ -37,6 +37,7 @@ year="2015",
 publisher="Springer International Publishing",
 address="Cham",
 pages="154--158",
+doi = {10.1007/978-3-319-17966-7_21},
 abstract="We present WebVOWL, a responsive web application for the visualization of ontologies. It implements the Visual Notation for OWL Ontologies (VOWL) and is entirely based on open web standards. The visualizations are automatically generated from JSON files, into which the ontologies need to be converted. An exemplary OWL2VOWL converter implemented in Java and based on the OWL API is currently used for this purpose. The ontologies are rendered in a force-directed graph layout according to the VOWL specification. Interaction techniques allow to explore the ontologies and customize their visualizations.",
 isbn="978-3-319-17966-7"
 }
