### 2024-07-10 cb62e3b
paper cut word count V2
