+++ JOSSpaper/paper.md
@@ -58,22 +58,20 @@ editor_options:
 
 # Summary
 
-The bilingual `FAIRmaterials` package is a robust and user-friendly tool designed to simplify the creation and visualization of materials and data science ontologies.
-`FAIRmaterials`, available in the Python and R languages, addresses the complexities and technical barriers associated with traditional ontology editors based on manual user input such as `Protege` [@Protege].
-`FAIRmaterials` offers an intuitive workflow and easy-to-use templates, making it accessible to users both experienced and inexperienced with ontology creation.
+The bilingual `FAIRmaterials` package simplifies the creation and visualization of materials and data science ontologies.
+`FAIRmaterials`, available in the Python and R languages, addresses the complexities associated with traditional ontology editors based on manual user input such as `Protege` [@Protege] with an intuitive workflow and easy-to-use templates, making it accessible to users both experienced and inexperienced with ontologies.
 
-The core feature of the `FAIRmaterials` package is its ability to programatically convert simple and structured CSV inputs into rich, well-defined ontologies.
+The `FAIRmaterials` package is its ability to programatically convert simple and structured CSV inputs into rich, well-defined ontologies.
 This capability is designed to support the findability, accessibility, interoperability, and reusability (FAIR) [@wilkinsonFAIRGuidingPrinciples2016] of research data and serve as a tool in the process of data FAIRification.
 
-Designed with the end-user in mind, `FAIRmaterials` facilitates a comprehensive yet entry-level experience in ontology creation.
 Its additional features, such as automated ontology merging, static visualizations, and comprehensive documentation for outputs extend its utility, making it a valuable tool for any researcher engaged in knowledge management.
 
 # Statement of need
 
 `Protege` is currently the most widely-used open-source tool for ontology creation and development.
 Its main capabilities include manually creating and editing ontological terms and relationships, visualizing ontologies, checking the logical consistency of ontologies, and querying ontologies for specific information.
-While Protege's extensive functionality even facilitates plugins, the complexity of the interface is a barrier for those who have little experience with ontology creation.
-This complexity prevents certain researchers from creating and integrating ontologies with their own datasets entirely.
+Unfortunately, the complexity of the interface is a barrier for those who have little experience with ontology creation.
+This complexity prevents many researchers from creating and integrating ontologies with their own datasets entirely.
 Therefore, there is a need for a tool that can create ontologies with an interface that is easily understandable and provides ample documentation on how to use it.
 `FAIRmaterials` seeks to lower the barrier of entry for scientists entering the world of ontology development and evolution.
 The package provides a baseline CSV ontology template with built-in and easy-to-follow instructions on how to design an ontology which can be found [here](https://docs.google.com/spreadsheets/d/1k7gFRc1Yslc-m65aWfFCxqk5UtrKZl9c3GyEFQvLSFU/edit?gid=1078491905#gid=1078491905).
@@ -96,36 +94,32 @@ The ontology is serialized into two syntaxes because of the unique advantages th
 Determining the correctness of an ontology is difficult if its representation is in a textual format.
 For this reason, the package outputs a visualization in both the R and Python versions.
 The optional Python flag include_graph_valuetype can be used to include value type nodes in the output visualization.
-The visualization is generated using the `Graphviz` [@GraphvizPython] software in the Python version and `DiagrammeR` [@DiagrammeR] in R. Both outputs are modeled after the popular `WebVOWL` [@Webvowl] ontology visualization tool to make it easier for users to inherently understand the color schema and format
[…]
