+++ Python/FAIRmaterials/FAIRmaterials.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+FAIRmaterials = FAIRmaterials.__main__:main
