+++ Python/FAIRmaterials/FAIRmaterials.egg-info/SOURCES.txt
@@ -1,0 +1,19 @@
+LICENSE.txt
+MANIFEST.in
+README.md
+setup.py
+FAIRmaterials/__init__.py
+FAIRmaterials/__main__.py
+FAIRmaterials/fair_sheet_parser.py
+FAIRmaterials/rdflib_graph_merger.py
+FAIRmaterials/rdflib_graph_saver.py
+FAIRmaterials.egg-info/PKG-INFO
+FAIRmaterials.egg-info/SOURCES.txt
+FAIRmaterials.egg-info/dependency_links.txt
+FAIRmaterials.egg-info/entry_points.txt
+FAIRmaterials.egg-info/requires.txt
+FAIRmaterials.egg-info/top_level.txt
+tests/__init__.py
+tests/fair_sheet_parser_test.py
+tests/rdflib_graph_merger_test.py
+tests/rdflib_graph_saver_test.py
