+++ Python/FAIRmaterials/FAIRmaterials.egg-info/requires.txt
@@ -1,0 +1,9 @@
+rdflib>=7.0.0
+pylode>=3.1.4
+matplotlib>=3.6.2
+numpy>=1.9.3
+graphviz>=0.20.1
+
+[dev]
+pytest
+pytest-cov

[…]
