### 2024-08-25 #6951 @berquist
> Hi @berquist, I am reaching out to invite you to review a manuscript titled "FAIRmaterials: Ontology Tools with Data FAIRification in Development," currently under consideration at the Journal of Open Source Software (JOSS). Your insights and feedback on this work would be valuable and much appreciated. To help with the review, we have [a reviewers' checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist) with useful guidelines. Thank you in advance for considering this invitation. Ana Trisovic, JOSS Topic Editor

Yes, I would be happy to review this. Any additional reviewers should be stronger than me in R, though :sweat_smile:
