### 2024-09-27 05eabd1
remove Python build products and venv
