+++ Python/FAIRmaterials/FAIRmaterials.egg-info/top_level.txt
@@ -1,2 +1,0 @@
-FAIRmaterials
-tests

[…]
