### 2024-10-05 #7287 @berquist
> @berquist I completed the merge. Are they any other actions I need to take for now?

Thank you for the poke.  There's nothing with the code, so I can continue with my review.  I do have some general questions.

For authorship, there are a number of authors on the paper, but according to Git, you are the only person who has contributed to the code.  Could you elaborate on this?

In the absence of running editorialbot (maybe there is a command for this) here are new LOC counts:
```
$ cloc .
      54 text files.
      35 unique files.
      64 files ignored.

Language                     files          blank        comment           code
HTML                             1            338              5           1607
SVG                              2              0            299           1483
R                                4            169            298            797
Python                           9            212            287            684
Markdown                         4            112              0            271
CSV                             10              0              0            145
TeX                              1              8              0             96
Rmd                              3            487           3528             33
Text                             1              4              0              5
SUM:                            35           1330           4417           5121
```
I also ran it on just the Python implementation, so no tests, and came up with
```
Python                           5            172            276            425
```
I don't necessarily view this as a measure of novelty, particular when there's an obvious need for the software, but in reading the paper, it's missing references to prior work in the area.  I've definitely seen tools that take ontologies defined in CSV or Excel workbooks and convert them to RDF.  The paper states that the tooling is for materials and data science ontologies, but aside from this initial mention, there are no references to the existing and surprisingly rich history of ontologies in materials science.  I've found at least https://link.springer.com/article/10.1557/s43579-024-00616-6 and https://link.springer.com/article/10.1557/s43580-024-00874-5 that claim usage of FAIRmaterials, so they need to be added.

There are some other criteria that may not be met under "substantial scholarly criteria", particularly age of software, though I get the impression it was simply not version controlled until recently. @atrisovic
