### 2024-10-05 #7287 @Jonathan-E-Gordon
@berquist Great thank you! I will look into adding the references.

In terms of the "substantial scholarly criteria" and the authorship you are correct we do not work in github. Due to lab policy, we developed the software in a private Bitbucket repository, which has accumulated over 400 commits since 2021. In January 2024, we initiated a redesign of the software, resulting in 110 commits from 9 authors, representing approximately 7 person-months of work. I can send the whole commit history if that is helpful. This software will also be cited in a Scientific Data paper currently under review, titled “Materials Data Science Ontology (MDS-Onto): Unifying Domain Knowledge in Materials and Applied Data Science”.
