### 2024-11-02 cwru-sdle/FAIRmaterials#2 @berquist
Python tests are not passing
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7287).

I installed the latest version of the code and the additional test dependencies:
```bash
python -m venv ./venv
source ./venv/bin/activate
python -m pip install -e .
python -m pip install pytest-cov
```
then ran the tests:
```bash
python -m pytest --cov=FAIRmaterials --cov-report=html
```
and got the following output:
```
==================================================================================== test session starts =====================================================================================
rootdir: /home/eric/data/manuscripts_review/joss_7287_fairmaterials/FAIRmaterials/Python/FAIRmaterials
plugins: anyio-4.6.2.post1, cov-6.0.0
collected 11 items

tests/fair_sheet_parser_test.py F                                                                                                                                                      [  9%]
tests/rdflib_graph_merger_test.py .F                                                                                                                                                   [ 27%]
tests/rdflib_graph_saver_test.py FF.F.F.F                                                                                                                                              [100%]

========================================================================================== FAILURES ==========================================================================================
_________________________________________________________________________________________ test_init __________________________________________________________________________________________

create_test_files = PosixPath('/tmp/pytest-of-eric/pytest-25/test_init0')

    def test_init(create_test_files):
        # Arrange
        ontology_sheet_folder = create_test_files
        include_graph_valuetype = False
        rdflib_graph = rdfGraph()
        graphviz_graph = graphviz.Digraph(strict=False)
        add_external_onto_info = False

        # Act
>       parser = FairSheetParser(ontology_sheet_folder, include_graph_valuetype, rdflib_graph, graphviz_graph, add_external_onto_info)

tests/fair_sheet_parser_test.py:131:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <FAIRmaterials.fair_sheet_parser.FairSheetParser object at 0x779ec02adbb0>, folder_path = PosixPath('/tmp/pytest-of-eric/pytest-25/test_init0'), include_graph_valuetype = False
rdflib_graph = <Graph identifier=Nd33d759ab10a40919da65b45aeb71990 (<class 'rdflib.graph.Graph'>)>, graphviz_graph = <graphviz.graphs.Digraph object at 0x779ec047a6c0>
add_external_onto_info = False

    def __init__(self, folder_path, include_graph_valuetype, rdflib_graph, graphviz_graph, add_external_onto_info):
        """
        Initializes the FairSheetParser object with the provided ontology sheet folder and populates the RDFLib graph and Graphviz PNG using the information provided in these sheets.

        Args:
            folder_path (str): The folder containing ontology CSV files.
            include_graph_valuetype (bool): Flag to include valuetype and unit edges in the Graphviz png.
            rdflib_graph (rdflib.Graph): An empty rdflib graph
            graphviz_graph (graphviz.Digraph): An empty graphviz graph
        """
>       ontology_info_path = next(iter(glob.glob(folder_path + "*- OntologyInfo.csv")))
E       TypeError: unsupported operand type(s) for +: 'PosixPath' and 'str'

FAIRmaterials/fair_sheet_parser.py:47: TypeError
________________________________________________________________________________ test_add_ontology_ownership _________________________________________________________________________________

ontology_one = <Graph identifier=Nca4096305b5c4bb2a14e0b079369a783 (<class 'rdflib.graph.Graph'>)>

    def test_add_ontology_ownership(ontology_one):
        base_uri = "http://example.org/"
        ontology_title = "TestOntology"
        ontology_version = "1.0"
        ontology_description = "This is an example description"

        merged_graph = RDFLibGraphMerger.add_ontology_ownership(ontology_one, base_uri, ontology_title, ontology_version, ontology_description)

        ontology_namespace = base_uri + "Ontology"

        # Check if the ontology metadata is added
>       assert (ontology_namespace, RDF.type, OWL.Ontology) in merged_graph
E       AssertionError: assert ('http://example.org/Ontology', rdflib.term.URIRef('http://www.w3.org/1999/02/22-rdf-syntax-ns#type'), rdflib.term.URIRef('http://www.w3.org/2002/07/owl#Ontology')) in <Graph identifier=Nca4096305b5c4bb2a14e0b079369a783 (<class 'rdflib.graph.Graph'>)>

tests/rdflib_graph_merger_test.py:42: AssertionError
______________________________________________________________________________ test_save_rdflib_graph_owl_valid ______________________________________________________________________________

saver = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec02aed50>

    def test_save_rdflib_graph_owl_valid(saver):
        file_path = saver.save_rdflib_graph_owl()
        assert os.path.exists(file_path)
>       assert file_path.endswith(".ttl")
E       AssertionError: assert False
E        +  where False = <built-in method endswith of str object at 0x779ec0339d40>('.ttl')
E        +    where <built-in method endswith of str object at 0x779ec0339d40> = 'test_ontology_output/test_ontology.owl'.endswith

tests/rdflib_graph_saver_test.py:50: AssertionError
______________________________________________________________________________ test_save_rdflib_graph_owl_empty ______________________________________________________________________________

empty_saver = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec0385010>

    def test_save_rdflib_graph_owl_empty(empty_saver):
>       with pytest.raises(ValueError):
E       Failed: DID NOT RAISE <class 'ValueError'>

tests/rdflib_graph_saver_test.py:56: Failed
____________________________________________________________________________ test_save_rdflib_graph_jsonld_empty _____________________________________________________________________________

empty_saver = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec19494f0>

    def test_save_rdflib_graph_jsonld_empty(empty_saver):
>       with pytest.raises(ValueError):
E       Failed: DID NOT RAISE <class 'ValueError'>

tests/rdflib_graph_saver_test.py:68: Failed
____________________________________________________________________________ test_save_graphviz_graph_valid_empty ____________________________________________________________________________

empty_saver = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec1949ac0>

    def test_save_graphviz_graph_valid_empty(empty_saver):
        with pytest.raises(gv_backend.ExecutableNotFound):
>           empty_saver.save_graphviz_graph()

tests/rdflib_graph_saver_test.py:79:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec1949ac0>

    def save_graphviz_graph(self):
        """
        Saves the Graphviz graph to a PNG file.

        Raises:
            graphviz.exceptions.GraphvizError: If the graph is empty and cannot be printed.
        """
        output_folder = f"{self.__ontology_name}_output"

        if not os.path.exists(output_folder):
            os.makedirs(output_folder)

        if len(self.__graphviz_graph.body) == 0:
>           raise graphviz.exceptions.GraphvizError("The graph is empty and cannot be printed")
E           AttributeError: module 'graphviz.exceptions' has no attribute 'GraphvizError'

FAIRmaterials/rdflib_graph_saver.py:88: AttributeError
______________________________________________________________________________ test_generate_pylode_html_empty _______________________________________________________________________________

empty_saver = <FAIRmaterials.rdflib_graph_saver.RDFLibGraphSaver object at 0x779ec0327920>

    def test_generate_pylode_html_empty(empty_saver):
        with pytest.raises(ValueError):
>           empty_saver.generate_pylode_html()

tests/rdflib_graph_saver_test.py:91:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
FAIRmaterials/rdflib_graph_saver.py:105: in generate_pylode_html
    op = VocPub(ontology=temp_file_path)
venv/lib/python3.12/site-packages/pylode/profiles/vocpub.py:121: in __init__
    self.ns = get_ns(self.ont)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

ont = <Graph identifier=N02c3c4edc5654a56857109adbea1cbb6 (<class 'rdflib.graph.Graph'>)>

    def get_ns(ont: Graph) -> Tuple[str, str]:
        """Gets the default Namespace for the given graph (ontology)"""
        # if this ontology declares a preferred URI, use that
        pref_iri = None
        for s_, o in ont.subject_objects(predicate=VANN.preferredNamespaceUri):
            pref_iri = str(o)

        pref_prefix = None
        for s_, o in ont.subject_objects(predicate=VANN.preferredNamespacePrefix):
            pref_prefix = str(o)
        if pref_prefix is None:
            pref_prefix = ""

        if pref_iri is not None:
            return pref_prefix, pref_iri

        # if not, try the URI of the main object, compared to all prefixes
        else:
            default_iri = None

            for s_ in chain(
                ont.subjects(predicate=RDF.type, object=OWL.Ontology),
                ont.subjects(predicate=RDF.type, object=SKOS.ConceptScheme),
                ont.subjects(predicate=RDF.type, object=PROF.Profile),
            ):
                default_iri = str(s_)

            if default_iri is not None:
                prefix = ont.compute_qname(default_iri, True)[0]
                if prefix is not None:
                    return prefix, default_iri
            else:
                # can't find either a declared or default namespace
                # so we have an error
>               raise Exception(
                    "pyLODE can't detect a URI for an owl:Ontology, "
                    "a skos:ConceptScheme or a prof:Profile"
                )
E               Exception: pyLODE can't detect a URI for an owl:Ontology, a skos:ConceptScheme or a prof:Profile

venv/lib/python3.12/site-packages/pylode/utils.py:105: Exception
====================================================================================== warnings summary ======================================================================================
tests/rdflib_graph_saver_test.py::test_generate_pylode_html_valid
  /home/eric/data/manuscripts_review/joss_7287_fairmaterials/FAIRmaterials/Python/FAIRmaterials/venv/lib/python3.12/site-packages/dominate/dom_tag.py:74: DeprecationWarning: There is no current event loop
    if get_event_loop().is_running():

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html

---------- coverage: platform linux, python 3.12.7-final-0 -----------
Coverage HTML written to dir htmlcov

================================================================================== short test summary info ===================================================================================
FAILED tests/fair_sheet_parser_test.py::test_init - TypeError: unsupported operand type(s) for +: 'PosixPath' and 'str'
FAILED tests/rdflib_graph_merger_test.py::test_add_ontology_ownership - AssertionError: assert ('http://example.org/Ontology', rdflib.term.URIRef('http://www.w3.org/1999/02/22-rdf-syntax-ns#type'), rdflib.term.URIRef('http://www.w3.org/2002/07/owl#Ontology'...
FAILED tests/rdflib_graph_saver_test.py::test_save_rdflib_graph_owl_valid - AssertionError: assert False
FAILED tests/rdflib_graph_saver_test.py::test_save_rdflib_graph_owl_empty - Failed: DID NOT RAISE <class 'ValueError'>
FAILED tests/rdflib_graph_saver_test.py::test_save_rdflib_graph_jsonld_empty - Failed: DID NOT RAISE <class 'ValueError'>
FAILED tests/rdflib_graph_saver_test.py::test_save_graphviz_graph_valid_empty - AttributeError: module 'graphviz.exceptions' has no attribute 'GraphvizError'
FAILED tests/rdflib_graph_saver_test.py::test_generate_pylode_html_empty - Exception: pyLODE can't detect a URI for an owl:Ontology, a skos:ConceptScheme or a prof:Profile
=========================================================================== 7 failed, 4 passed, 1 warning in 1.64s ===========================================================================
```

---
You don't need to check coverage to look at the failures, but I wanted to look at it later.
