### 2024-11-02 #7287 @berquist
Apologies for letting my review slip.

I don't think JOSS has a policy about making private code public, but since the review should be in the open, I wouldn't request to see `git log > git_log.txt` unless you were willing to post it publicly as an attachment in this comment thread. At this point I'm still looking for editor guidance.

I've created an issue on the repo about the failing Python tests and a PR for discussing how the READMEs should be handled, since there is no other project documentation or examples.

Although the READMEs do cover an example of using the code from both Python and R, it would be much better to have this fully-worked example present in the repository.  I think it's present in the R package but not the Python package.
