### 2024-11-22 #7287 @haydencal
Hello @berquist! I am so sorry for the extensive delay on this, but my name is Hayden and I am an undergraduate working with the Python package. I am in the process of debugging the package so that it can pass the appropriate tests, and will also look into the issue with the README files ASAP. Attached below is the git log for our repository in Bitbucket so you can see our development history. I will continue working with the problems with the Python package and update you accordingly as appropriate.

[git_log.txt](https://github.com/user-attachments/files/17854095/git_log.txt)
