### 2024-11-25 #7287 @haydencal
Sounds good, I can talk to the team working with the R code as well if
there is something in particular that comes up.

On Mon, Nov 25, 2024 at 12:11 AM Eric Berquist ***@***.***>
wrote:

> Great, thank you. Let me know if there is anything I can help with in the
> meantime; the one thing I haven't done yet is look at the R code.
>
