### 2024-12-13 d4f4d2e
python test fixes
