### 2025-01-20 #7287 @atrisovic
Hi @berquist and @emanueledelsozzo 👋

I hope you're doing well. Just checking in to see how you're getting along with the review of "FAIRmaterials: Ontology Tools with Data FAIRification in Development". If you have any questions or need further information, please don't hesitate to let me know.
Looking forward to your insights!
Ana
