### 2025-03-11 #7287 @berquist
@Jonathan-E-Gordon Has any progress been made on https://github.com/cwru-sdle/FAIRmaterials/issues/2?  Same question about the references in the paper.  Without both of these done, I can't recommend this for publication.
