### 2025-03-14 ca1cb57
Updated citations for use in literature
