+++ JOSSpaper/paper.md
@@ -7,9 +7,9 @@ tags:
 - FAIR
 date: "26 June 2024"
 output:
+  pdf_document: default
   html_document:
     df_print: paged
-  pdf_document: default
 authors:
 - name: Alexander Harding Bradley
   affiliation: 2, 3
@@ -74,7 +74,8 @@ Unfortunately, the complexity of the interface is a barrier for those who have l
 This complexity prevents many researchers from creating and integrating ontologies with their own datasets entirely.
 Therefore, there is a need for a tool that can create ontologies with an interface that is easily understandable and provides ample documentation on how to use it.
 `FAIRmaterials` seeks to lower the barrier of entry for scientists entering the world of ontology development and evolution.
-The package provides a baseline CSV ontology template with built-in and easy-to-follow instructions on how to design an ontology which can be found [here](https://docs.google.com/spreadsheets/d/1k7gFRc1Yslc-m65aWfFCxqk5UtrKZl9c3GyEFQvLSFU/edit?gid=1078491905#gid=1078491905).
+The package provides a baseline CSV ontology template with built-in and easy-to-follow instructions on how to design an ontology which can be found [here](https://docs.google.com/spreadsheets/d/1k7gFRc1Yslc-m65aWfFCxqk5UtrKZl9c3GyEFQvLSFU/edit?gid=1078491905#gid=1078491905). 
+Preliminary applications of the package in additive manufacturing [@FAIRcite1] suggest that `FAIRmaterials` offers valuable capabilities for researchers.
 
 ![Empty Variable CSV Template Sheet for the `FAIRmaterials` Package. The CSV template sheet includes specific instructions on how to fill out every row to correctly generate the ontology. Template is split in half for readability.\label{fig:Variable_Sheet_Template_CSV}](Variable_Sheet_Template_CSV.png){width="100%"}
 
