+++ JOSSpaper/paper.bib
@@ -102,3 +102,15 @@ isbn="978-3-319-17966-7"
   howpublished = {\url{https://cran.r-project.org}},
   note         = {Accessed: 2024-06-18}
 }
+
+@article{FAIRcite1,
+  author = {Hernandez, K.J. and Barcelos, E.I. and Jimenez, J.C. and others},
+  title = {A data integration framework of additive manufacturing based on FAIR principles},
+  journal = {MRS Advances},
+  volume = {9},
+  pages = {844--851},
+  year = {2024},
+  doi = {10.1557/s43580-024-00874-5},
+  url = {https://doi.org/10.1557/s43580-024-00874-5},
+  note = {Received: 30 January 2024; Accepted: 07 May 2024; Published: 28 May 2024; Issue Date: July 2024}
+}
