### 2025-03-19 0f6d1f5
Updated citations for use in literature
