+++ JOSSpaper/paper.bib
@@ -114,3 +114,13 @@ isbn="978-3-319-17966-7"
   url = {https://doi.org/10.1557/s43580-024-00874-5},
   note = {Received: 30 January 2024; Accepted: 07 May 2024; Published: 28 May 2024; Issue Date: July 2024}
 }
+
+@inproceedings{Nihar2021,
+  author    = {A. Nihar and others},
+  title     = {Toward Findable, Accessible, Interoperable and Reusable (FAIR) Photovoltaic System Time Series Data},
+  booktitle = {2021 IEEE 48th Photovoltaic Specialists Conference (PVSC)},
+  address   = {Fort Lauderdale, FL, USA},
+  year      = {2021},
+  pages     = {1701--1706},
+  doi       = {10.1109/PVSC43889.2021.9518782}
+}
