### 2025-05-30 #7287 @vandttran
Hi @berquist, all issues you mentioned should be resolved. I've made sure all the README.md files point to the github folder containing the template. This folder now contains CSVs and one xlsx file. I've also added a worked example using the XRD Sample Ontology. Thanks!
