### 2026-01-20 #7287 @atrisovic
Great, thank you! I am also back from a longer break and will get back to this asap!

PS: Any thoughts on 10.1016/j.artmed.2017.07.002 may be a valid DOI for title: Owlready: Ontology-oriented programming in Python?
