### 2026-01-20 #7287 @atrisovic
@vandttran Can you make Figures 6, 7, and 8 landscape (horizontal)? They are hard to see in the current orientation.
