### 2026-01-21 #7287 @vandttran
Hi @atrisovic, I've increased the font size in the figure. I think that's the best I can do for visibility.
