### 2026-01-24 #7287 @danielskatz
As track editor, I'll work on this later today
