### 2026-01-24 #7287 @danielskatz
👋 @berquist - I note you didn't check the community guidelines item in your checklist, and there's an open issue - https://github.com/cwru-sdle/FAIRmaterials/issues/21

Should this item now be checked off and the open issue closed?
