### 2026-01-24 #7287 @danielskatz
👋 @vandttran & @Alexhb02 - I've suggested a few small changes in https://github.com/cwru-sdle/FAIRmaterials/pull/28 - please merge this or let me know what you disagree with, then we can proceed
