### 2026-01-25 #7287 @vandttran
I've merged the changes @danielskatz. Thank you for the fixes.
