### 2026-01-25 #7287 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
