### 2026-01-26 #7287 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
