### 2026-01-26 #7287 @danielskatz
Congratulations to @Alexhb02 (Alexander Harding Bradley) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you.
